+++ Unity-Inventory-System/Assets/InventorySystem/Runtime/Scripts/Core/ViewModels/Inventory/ItemViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace Assets.Scripts.Core.ViewModels
+﻿namespace InventorySystem.Runtime.Scripts.Core.ViewModels.Inventory
 {
 	public class ItemViewModel
 	{
