+++ Unity-Inventory-System/Assets/InventorySystem/Runtime/Scripts/Installers/Inventory/HotBarInstaller.cs
@@ -1,7 +1,7 @@
-﻿using UnityInventorySystem.Inventory;
+﻿using InventorySystem.Runtime.Scripts.Inventory;
 using Zenject;
 
-namespace UnityInventorySystem.Installers
+namespace InventorySystem.Runtime.Scripts.Installers.Inventory
 {
 	public class HotBarInstaller : Installer<HotBarInstaller>
 	{
