+++ Unity-Inventory-System/Assets/InventorySystem/Runtime/Scripts/Installers/Inventory/ItemInstaller.cs
@@ -1,10 +1,9 @@
-﻿using Assets.Scripts.Core.ViewModels;
+﻿using InventorySystem.Runtime.Scripts.Core.ViewModels.Inventory;
 using InventorySystem.Runtime.Scripts.Inventory.Item;
 using InventorySystem.Runtime.Scripts.Models;
-using UnityInventorySystem.Inventory;
 using Zenject;
 
-namespace UnityInventorySystem.Installers
+namespace InventorySystem.Runtime.Scripts.Installers.Inventory
 {
 	public class ItemInstaller : Installer<ItemInstaller>
 	{
