+++ Unity-Inventory-System/Assets/InventorySystem/Runtime/Scripts/Installers/Inventory/SlotsContainerInstaller.cs
@@ -1,8 +1,7 @@
 ﻿using InventorySystem.Runtime.Scripts.Inventory.Slot;
-using UnityInventorySystem.Inventory;
 using Zenject;
 
-namespace UnityInventorySystem.Installers
+namespace InventorySystem.Runtime.Scripts.Installers.Inventory
 {
 	public class SlotsContainerInstaller : Installer<SlotsContainerInstaller>
 	{
