+++ Unity-Inventory-System/Assets/InventorySystem/Runtime/Scripts/Inventory/HotBarFacade.cs
@@ -1,6 +1,6 @@
 ﻿using Zenject;
 
-namespace UnityInventorySystem.Inventory
+namespace InventorySystem.Runtime.Scripts.Inventory
 {
 	public class HotBarFacade
 	{
