+++ Unity-Inventory-System/Assets/InventorySystem/Samples/Scripts/Installers/MainScene/MainInventorySOInstaller.cs
@@ -1,16 +1,17 @@
-using System;
 using UnityEngine;
-using UnityInventorySystem.Installers;
 using Zenject;
 
-[CreateAssetMenu(fileName = "MainInventorySOInstaller", menuName = "SO/Installers/MainInventorySOInstaller")]
-public class MainInventorySOInstaller : ScriptableObjectInstaller<MainInventorySOInstaller>
+namespace InventorySystem.Samples.Scripts.Installers.MainScene
 {
-    #pragma warning disable 0649
-    [SerializeField] private MainInventoryMonoInstaller.Settings _mainInventoryMonoInstallerSettings;
-    #pragma warning restore 0649
-    public override void InstallBindings()
-    {
-        Container.BindInstance(_mainInventoryMonoInstallerSettings).IfNotBound();
-    }
+	[CreateAssetMenu(fileName = "MainInventorySOInstaller", menuName = "SO/Installers/MainInventorySOInstaller")]
+	public class MainInventorySOInstaller : ScriptableObjectInstaller<MainInventorySOInstaller>
+	{
+#pragma warning disable 0649
+		[SerializeField] private MainInventoryMonoInstaller.Settings _mainInventoryMonoInstallerSettings;
+#pragma warning restore 0649
+		public override void InstallBindings()
+		{
+			Container.BindInstance(_mainInventoryMonoInstallerSettings).IfNotBound();
+		}
+	}
 }
