+++ Unity-Inventory-System/Assets/InventorySystem/Samples/Scripts/InventoryItemsSO.cs
@@ -1,7 +1,7 @@
 ﻿using Malee.List;
 using UnityEngine;
 
-namespace UnityInventorySystem
+namespace InventorySystem.Samples.Scripts
 {
 	[CreateAssetMenu(fileName = "InventorySlots", menuName = "SO/InventoryItemsSO")]
 	public class InventoryItemsSo : ScriptableObject
